Refactor to ES6
