+++ .babelrc
@@ -1,0 +1,7 @@
+{
+  "presets": [
+    "react",
+    "es2015",
+    "stage-3"
+  ]
+}
